+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Print error if variable is missing for Jinja2 (optional variables now need to stricly check with `is defined`)
+
 ## [0.2.2] - 2025-04-01
 
 ### Added