+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.3] - 2025-04-02
+
 ### Changed
 
 - Validate variables for sub-findings if the corresponding template can be found.
@@ -312,7 +314,8 @@ We registered a dummy package to PyPI to test the publishing. Therefore a versio
 Initial version
 
 
-[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.2.2...HEAD
+[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.2.3...HEAD
+[0.2.3]: https://github.com/s3r3t0/sereto/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/s3r3t0/sereto/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/s3r3t0/sereto/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/s3r3t0/sereto/compare/v0.1.1...v0.2.0